requirements.txt
lightgbm>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
ta>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
textblob>=0.17.0
nltk>=3.8.0
streamlit>=1.29.0
pyyaml>=6.0.0
rich>=13.7.0
mlflow>=2.8.0
yfinance>=0.2.50
matplotlib>=3.8.0
shap>=0.44.0
